Feat: Prepare play (#16)
This commit implements the preparation of a play. This consists essentially in being able to drag and drop the tiles either in the board or in the deck in a different position.

Moving a tile and putting it anywhere of the allowed places will trigger an update in the database so that next time you reload the session you can see the tile in the position you left it.

It also implements optimistic locks so that, if the user has multiple tabs open with the same match and stale data, he won't be able to update his hand forcing him to reload the page.