Feat: Plays (2/?) - WordScanner (#20)
Done in this commit:

Maptrix module

Maptrix stands for Map and Matrix. It basically has different methods to transform items belonging to the board to a map where the coordinates will be the key. The key will be a tuple including based on the row and the column.

This module is really helpful for PlayIntegrityValidator and WordScanner as they are constantly checking what's in each position.

PlayValidator is now PlayIntegrityValidator

The name has been changed with something that can give more clues about what it does. In case of success the validate/2 function will return {:ok, alignment} instead of just :ok.

The alignment is helpful as it will make much easier the work for the WordScanner narrowing down the words.

Also, originally the validate/2 method received %Match{} and %Player{} structs as arguments but they have been replaced with a Maptrix of the board and the list of the playing tiles of the player.

WordScanner

WordScanner and scan/3 takes a maptrix of the already played tiles, the tiles a player is proposing and the orientation previously figured out by the PlayIntegrityValidator and returns a list of words that have been formed as a result of dropping the player tiles in the board.

These words are merely adjacent tiles. They aren't necessarily existing entries in the lexicon. This will be verified in a different step done by a different module. This way we keep the responsibilities separated properly.

positions attribute has been a last minute addition to the Word embedded schema. This represent a list of all the coordinates that are involved in building a Word. It will become very useful later on when the board component shows highlighted the play in the board