Feat: Plays (4/4) - Play creation transaction (#24)
This big commit wraps up the Feat: Plays series.

In this one we put all the previous code to work inside Games.create_play/2 function.

Games.create_play/2

This function receives the match and the player that is submitting a play and does the following:

Validates the tiles are properly positioned
Scan the words builds on the board
Checks all the words built are actual entries in the match lexicon
Starts a transaction in 3 steps
Creates the play record
Updates the match increasing the turn and drawing the tiles required by the player after his play.
Updates the player adding the points of the plays to the score and the fresh new tiles from the bag.
GameBoardComponent

The live component has been updated to connect the submit button click to the corresponding action described above.

After succeeding with a play it will be broadcasted to all the players which will see the board and the scores updated.

Errors will be shown in flash notifications in case of failure.

ScrabblexWeb.Flash

This module has been added to make it possible to trigger flash notifications from live components. Kudos to @sevenseacat for her solution.