Feat: Exchange tiles (#31)
Tile exchange with the bag has been implemented in a similar fashion as the wildcard editor.

On player's turn an "exchange tiles" button will become visible. On clicking it a modal will be brought up where player can select a number of tiles to be exchanged with the ones that are still in the bag.