Feat:  Display winner when match is finished (#58)
This change adds a new modal that will pop up when the match status transitions to "finished". The modal will show the winner by detecting the player with the max amount of points.

The actions bar will become hidden once the status is transitioned to "finished" preventing players from submitting more plays.